fix: survive invalid json tmpfile on init()
Currently if you are missing a cache file during init() > meta >
activeLoader() > tryCache(), shared-store gracefully returns
Observable.empty() and gets on with life.  If the cache file happens to
be invalid, it dies, forcing you to restart your app to get a fresh
copy.  This change treats syntax errors the same as missing files (modulo a
debug line).